feat: responsive manage & add item
